PLANET-8181: Dynamically reload posts on filters selection
- https://github.com/greenpeace/planet4-master-theme/pull/3047#discussion_r3657970175
